李宇辉：添加超级管理员后台操作模块

+++ 20151104687_liyuhui/20151104687_liyuhui_LaboratoryCloudMonitoringSystem/src/main/java/com/lyh/springboot/mapper/RoleMapper.java
@@ -18,7 +18,7 @@ public interface RoleMapper {
 
     List<Role> selectByExample(RoleExample example);
 
-    Role selectByPrimaryKey(Integer id);
+    Role selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Role record, @Param("example") RoleExample example);
