+++ 20151104687_liyuhui/20151104687_liyuhui_LaboratoryCloudMonitoringSystem/src/main/java/com/lyh/springboot/pojo/UserRole.java
@@ -3,35 +3,35 @@ package com.lyh.springboot.pojo;
 import java.io.Serializable;
 
 public class UserRole implements Serializable {
-    private Integer id;
+    private Long id;
 
-    private Integer uid;
+    private Long uid;
 
-    private Integer rid;
+    private Long rid;
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public void setUid(Integer uid) {
+    public void setUid(Long uid) {
         this.uid = uid;
     }
 
-    public Integer getRid() {
+    public Long getRid() {
         return rid;
     }
 
-    public void setRid(Integer rid) {
+    public void setRid(Long rid) {
         this.rid = rid;
     }
 }
